fix: separate style code and change file routes (#8)
