+++ src/pages/Landing/Landing.tsx
@@ -1,6 +1,7 @@
-import { styled } from 'styled-components';
 import React from 'react';
 import { useNavigate } from 'react-router-dom';
+import MenuCard from '../../components/common/MenuCard/MenuCard';
+import { Layout, Wrapper } from './Landing.styled';
 
 const Landing = () => {
   const navigate = useNavigate();
@@ -12,14 +13,15 @@ const Landing = () => {
   return (
     <Layout>
       <button onClick={handleStartButton}> 시작</button>
+      <Wrapper>
+        <MenuCard />
+        <MenuCard />
+        <MenuCard />
+        <MenuCard />
+        <MenuCard />
+      </Wrapper>
     </Layout>
   );
 };
 
 export default Landing;
-
-const Layout = styled.div`
-  width: 100%;
-  height: 100%;
-  background-color: var(--main-orange);
-`;
