+++ src/router/AppRoutes.tsx
@@ -1,19 +1,19 @@
 import React from 'react';
 import { Routes, Route } from 'react-router-dom';
-import Landing from '../pages/Landing';
-import Location from '../pages/Location';
-import RandomList from '../pages/RandomList';
-import Result from '../pages/Result';
-import Vote from '../pages/Vote';
+import Landing from '../pages/Landing/Landing';
+import Location from '../pages/Location/Location';
+import RandomList from '../pages/RandomList/RandomList';
+import Result from '../pages/Result/Result';
+import Vote from '../pages/Vote/Vote';
 
 const AppRoutes = () => {
   return (
     <Routes>
-      <Route path='/' element={<Landing />} />
-      <Route path='/location' element={<Location />} />
-      <Route path='/random-menu' element={<RandomList />} />
-      <Route path='/random-menu/:id' element={<Vote />} />
-      <Route path='/random-menu/:id/result' element={<Result />} />
+      <Route path="/" element={<Landing />} />
+      <Route path="/location" element={<Location />} />
+      <Route path="/random-menu" element={<RandomList />} />
+      <Route path="/random-menu/:id" element={<Vote />} />
+      <Route path="/random-menu/:id/result" element={<Result />} />
     </Routes>
   );
 };
